Mask Gen for Re-ID dataset

+++ data_preparation.py
@@ -19,7 +19,7 @@ import PartAttGen
 import model
 
 
-def pipeline_span(image_root, mask_dl_ckpt, part_att_ckpt):
+def pipeline_span(Or_image_root, mask_dl_ckpt, part_att_ckpt, target_dir):
 #if __name__ == '__main__':
     torch.manual_seed(1234)
     torch.cuda.manual_seed(1234)
@@ -41,39 +41,55 @@ def pipeline_span(image_root, mask_dl_ckpt, part_att_ckpt):
     #parser.add_argument('--part_att_root', default='./PartAttMask', help='path to generated part attention mask')
     #-----parser.add_argument('--part_att_ckpt', default='./PartAttMask_ckpt', help='path to store part attention mask generator checkpoint')
     #-----args = parser.parse_args()
+    temp_dataset_root=Or_image_root+'/temp_data'
+    if not os.path.isdir(temp_dataset_root):
+        os.mkdir(temp_dataset_root)
+    
 
+    f_list=os.listdir(Or_image_root)
+    for i in f_list:
+        shutil.copytree(Or_image_root+'/'+i , temp_dataset_root+'/'+i)
+
+    image_root=temp_dataset_root
+    
     Folder_list=os.listdir(image_root)
     for k in Folder_list:
-        
-        for (root,dirs,files) in os.walk(image_root+'/'+k, topdown=True):
-            if len(dirs) == 0:
-                for i in files:
-                    shutil.copy(root+'/'+i, image_root+'/'+k+'/'+i)
-                print(root)
-                shutil.rmtree(root)
+        if k!='temp_data':
+            for (root,dirs,files) in os.walk(image_root+'/'+k, topdown=True):
+                if len(dirs) == 0:
+                    for i in files:
+                        shutil.copy(root+'/'+i, image_root+'/'+k+'/'+i)
+                    print(root)
+                    shutil.rmtree(root)
 
-        part_att_root = image_root+'/mask_generate_{}/part_attention'.format(k)
+    tar_dir_name=target_dir+'/attention_masks_gen'
+    if not os.path.isdir(tar_dir_name):
+        os.mkdir(tar_dir_name)
+
+    part_att_root = tar_dir_name
+
+    print("\n### Generate part attention mask ###")
+    checkpoint = os.path.join(part_att_ckpt, '10.ckpt')
+    PartAttGen.implement(image_root=image_root,
+                        mask_root= part_att_root,
+                        model= model.PartAtt_Generator().to(device),
+                        device= device,
+                        checkpoint= checkpoint)
 
 
-        print("\n### Generate part attention mask ###")
-        checkpoint = os.path.join(part_att_ckpt, '10.ckpt')
-        PartAttGen.implement(image_root=image_root,
-                            mask_root= part_att_root,
-                            model= model.PartAtt_Generator().to(device),
-                            device= device,
-                            checkpoint= checkpoint)
-
-
-        for (root,dirs,files) in os.walk(image_root+'/mask_generate_'+k+'/part_attention', topdown=True):
-            if len(dirs)==0:
-                for i in files:
-                    shutil.copy(root+'/'+i, image_root+'/mask_generate_'+k+'/part_attention'+'/'+i)
-                shutil.rmtree(root)
-
+        # for (root,dirs,files) in os.walk(image_root+'/mask_generate_'+k+'/part_attention', topdown=True):
+        #     if len(dirs)==0:
+        #         for i in files:
+        #             shutil.copy(root+'/'+i , image_root+'/mask_generate_'+k+'/part_attention'+'/'+i)
+        #         shutil.rmtree(root)
+    
+    shutil.rmtree(temp_dataset_root)
+    shutil.rmtree(part_att_root+'/temp_data')
     return 0
 
         
 
-pipeline_span("/home/fyp3/Desktop/Batch18/Re_ID/Dataset/re_id_weligama", "/home/fyp3/Desktop/Batch18/Re_ID/Dataset/mask_dl_chckpt/", "/home/fyp3/Desktop/Batch18/Re_ID/Dataset/part_attention_ckpt/")
+
+pipeline_span(Or_image_root="/home/fyp3/Desktop/Batch18/Re_ID/Dataset/re_id_weligama", mask_dl_ckpt="/home/fyp3/Desktop/Batch18/Re_ID/Dataset/mask_dl_chckpt/", part_att_ckpt="/home/fyp3/Desktop/Batch18/Re_ID/Dataset/part_attention_ckpt/" , target_dir="/home/fyp3/Desktop/Batch18/Re_ID/" )
